pyproject.toml
[project]
name = "ame-analylics"
version = "0.1.0"
description = ""
authors = [
    {name = "ezequiel.bernatene",email = "ezequiel.bernatene@ameol.it"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
